[nvim][shell][zsh] remove nvim cd hook
The cd hook was *extremely* slow. cd-ing to a directory without needing
to invoke the neovim host for any task means the hook was unnecessarily
triggered; I suspect that this is the usual case.

In reality, you don't really need to sync the shell's cwd with the
hosting neovim's cwd in order to achieve the functionality of aliases
like "e" and "tabe"; but we do it anyway, only lazily. This ends up
working as before, changing the host neovim's cwd to the shell's current
one, but only *if* and *precisely when* we invoke it for some reason.